RChain: renamed rFunctions files
