+++ server/modules/analysis/service/_rChain/rFunctions/arena.js
