+++ server/modules/analysis/service/_rChain/rFunctions/index.js
@@ -1,7 +1,7 @@
-export { NA, dfVar, merge, setVar, sysTime } from './rFunctionsUtility'
+export { NA, dfVar, merge, setVar, sysTime } from './utility'
 
-export { dbDisconnect, dbGetQuery, dbSendQuery, dbWriteTable, setConnection } from './rFunctionsDb'
+export { dbDisconnect, dbGetQuery, dbSendQuery, dbWriteTable, setConnection } from './db'
 
-export { arenaStartTime, arenaEndTime, arenaInfo, arenaPersistCalculationScript } from './rFunctionsArena'
+export { arenaStartTime, arenaEndTime, arenaInfo, arenaPersistCalculationScript } from './arena'
 
-export { arenaGet, arenaPost } from './rFunctionsArenaApi'
+export { arenaGet, arenaPost } from './arenaApi'
