+++ server/modules/analysis/service/_rChain/rFunctions/utility.js
